+++ src/lib/icons/icons.tsx
@@ -2,11 +2,15 @@
 import ArrowDown from "@mui/icons-material/KeyboardArrowDownRounded";
 import ArrowUp from "@mui/icons-material/KeyboardArrowUpRounded";
 
+// C
+import CloseIcon from '@mui/icons-material/CloseRounded';
+
 // S
 import ShareIcon from '@mui/icons-material/ShareRounded';
 
 export {
     ArrowDown,
     ArrowUp,
-    ShareIcon
+    ShareIcon,
+    CloseIcon
 }