+++ src/lib/table/table_interactive.tsx
@@ -1,13 +1,18 @@
 // React
 import React, { useState } from 'react';
 
-// Typography
+// Vieolo UI
+import Flex from '../layout/flex/flex';
 import TypographyParagraphMedium from '../typography/typography_paragraph_medium';
 import TypographyParagraphSmall from '../typography/typography_paragraph_small';
 import TypographyTitleSmall from '../typography/typography_title_small';
 
 // Types
 import { ColorOptionType } from '../private/types';
+import { toFixedFloat } from '@vieolo/parsers';
+import Input from '../form/input';
+import IconButton from '../button/icon_button';
+import { CloseIcon } from '../icons/icons';
 
 
 export type TableInteractiveCell = {
@@ -24,12 +29,23 @@ export type TableInteractiveCell = {
         span: number
     },
     numericalValue?: number,
-    background?: ColorOptionType,    
+    background?: ColorOptionType,
+    onTextEdit?: (newValue: string) => void
+}
+
+
+type HeaderFormatter = (s: string | number) => string;
+
+type EditCellData = {
+    rowIndex: number,
+    cellIndex: number,
+    initialText: string,
+    newText: string
 }
 
 
 export default function TableInteractive(props: {
-    headers: (string | {name: string, formatter: (s: string | number) => string})[],
+    headers: (string | { name: string, formatter: HeaderFormatter })[],
     headerSticky?: boolean,
     rows: TableInteractiveCell[][],
     selectedColor?: ColorOptionType
@@ -37,11 +53,14 @@ export default function TableInteractive(props: {
     width?: string,
     showBottomRow?: boolean,
     isDense?: boolean,
-    columnSelectedTotalFormat?: {[column: number]: (total: number) => string}
+    /** @deprecated */
+    columnSelectedTotalFormat?: { [column: number]: (total: number) => string }
 }): JSX.Element {
 
     let [selectedCells, setSelectedCells] = useState<string[]>([]);
+    let [selectedCellColumn, setSelectedCellColumn] = useState<number>(-1);
     let [numericTotal, setNumericTotal] = useState<number>(0);
+    let [editCellData, setEditCellData] = useState<EditCellData | null>(null)
 
     let style: React.CSSProperties = {}
 
@@ -49,22 +68,25 @@ export default function TableInteractive(props: {
         style.width = `${props.width}`;
     }
 
-    function toggleSelectCell(key: string, numbericValue: number, action: 'add' | 'remove-last' | 'reset' | 'only-entry') {
+    function toggleSelectCell(key: string, numbericValue: number, action: 'add' | 'remove-last' | 'reset' | 'only-entry', column: number) {
         switch (action) {
             case 'add':
                 setSelectedCells([...selectedCells, key])
                 setNumericTotal(numericTotal + numbericValue);
+                setSelectedCellColumn(column);
                 break;
             case 'remove-last':
                 setSelectedCells([...selectedCells].slice(0, -1))
                 break;
-            case 'reset': 
+            case 'reset':
                 setSelectedCells([]);
                 setNumericTotal(0);
+                setSelectedCellColumn(-1);
                 break;
-            case 'only-entry': 
+            case 'only-entry':
                 setSelectedCells([key]);
                 setNumericTotal(numbericValue);
+                setSelectedCellColumn(column);
                 break;
             default:
                 break;
@@ -74,7 +96,7 @@ export default function TableInteractive(props: {
     let cellClass = `vieolo-table-interactive__cell vieolo-table-interactive__cell--height-${props.isDense ? 'small' : 'medium'}`
 
     return <div className="vieolo-table-interactive" style={style}>
-        <div className={`vieolo-table-interactive__header-row ${props.headerSticky ? "position--sticky--top-0" : ""}`} style={{gridTemplateColumns: props.columnGrid}}>
+        <div className={`vieolo-table-interactive__header-row ${props.headerSticky ? "position--sticky--top-0 z-index--3" : ""}`} style={{ gridTemplateColumns: props.columnGrid }}>
             {
                 props.headers.map(h => {
                     let s = typeof h === 'string' ? h : h.name;
@@ -87,13 +109,30 @@ export default function TableInteractive(props: {
 
         {
             props.rows.map((row, i) => {
-                return <div className="vieolo-table-interactive__content-row" style={{gridTemplateColumns: props.columnGrid}} key={`row_${i}`}>
+                return <div className="vieolo-table-interactive__content-row" style={{ gridTemplateColumns: props.columnGrid }} key={`row_${i}`}>
                     {
                         row.map((r, ri) => {
                             let k = `${i}_${ri}`;
+
+                            if (editCellData && editCellData.cellIndex === ri && editCellData.rowIndex === i) {
+                                return <InputCell
+                                    cellClass={cellClass}
+                                    onChange={v => setEditCellData({ ...editCellData as any, newText: v })}
+                                    onSubmit={() => {
+                                        if (r.onTextEdit) {
+                                            r.onTextEdit(editCellData!.newText)
+                                            setEditCellData(null)
+                                        }
+                                    }}
+                                    value={editCellData.newText}
+                                    key={k + "_input_cell"}
+                                    onCancel={() => setEditCellData(null)}
+                                />
+                            }
+
                             let className = cellClass;
-                            
-                            if (r.onClick || r.selectable) {
+
+                            if (r.onClick || r.selectable || r.onTextEdit) {
                                 className += ' cursor--pointer nonselectable vieolo-table-interactive__cell--hover';
                             }
 
@@ -120,7 +159,7 @@ export default function TableInteractive(props: {
                                 style.right = '0';
                                 style.zIndex = 2;
                             }
-                            
+
                             let finalNode = r.value;
 
                             // Getting the header
@@ -128,26 +167,34 @@ export default function TableInteractive(props: {
                             // If the header contains the `formatter` function, the value of formatter function is displayed
                             let header = props.headers[ri];
 
-                            if (typeof r.value === 'string' || typeof r.value === 'number') finalNode = <TypographyParagraphMedium text={typeof header === 'string' ? r.value.toString() : header.formatter( r.numericalValue || r.value)} />;
-                            
+                            if (typeof r.value === 'string' || typeof r.value === 'number') finalNode = <TypographyParagraphMedium text={typeof header === 'string' ? r.value.toString() : header.formatter(r.numericalValue || r.value)} />;
 
-                            return <div 
+
+                            return <div
                                 className={className}
                                 key={k}
                                 onClick={() => {
                                     if (r.onClick) r.onClick(r.id);
+                                    else if (r.onTextEdit && typeof r.value === 'string') {
+                                        setEditCellData({
+                                            cellIndex: ri,
+                                            rowIndex: i,
+                                            initialText: r.value,
+                                            newText: r.value
+                                        })
+                                    }
                                 }}
                                 onMouseDown={e => {
                                     if (r.selectable && r.numericalValue) {
-                                        if (selectedCells.length === 0) toggleSelectCell(k, r.numericalValue, 'only-entry')
-                                        else toggleSelectCell(k, r.numericalValue, 'reset')
-                                    }                                    
+                                        if (selectedCells.length === 0) toggleSelectCell(k, r.numericalValue, 'only-entry', ri)
+                                        else toggleSelectCell(k, r.numericalValue, 'reset', ri)
+                                    }
                                 }}
-                                onMouseEnter={e => {                                    
-                                    if (r.selectable && e.buttons && r.numericalValue) {                    
-                                        if (selectedCells.includes(k)) toggleSelectCell(k, r.numericalValue, 'remove-last')
-                                        else toggleSelectCell(k, r.numericalValue, 'add')
-                                    }                                    
+                                onMouseEnter={e => {
+                                    if (r.selectable && e.buttons && r.numericalValue) {
+                                        if (selectedCells.includes(k)) toggleSelectCell(k, r.numericalValue, 'remove-last', ri)
+                                        else toggleSelectCell(k, r.numericalValue, 'add', ri)
+                                    }
                                 }}
                             >
                                 {
@@ -170,13 +217,79 @@ export default function TableInteractive(props: {
 
         {
             props.showBottomRow &&
-            <div className={`vieolo-table-interactive__bottom-row`}>
-                {
-                    (selectedCells.length > 0 && !isNaN(numericTotal)) &&
-                    <TypographyParagraphSmall text={numericTotal.toString()} />
-                }
-            </div>
+            <BottomRow
+                header={props.headers[selectedCellColumn]}
+                numericTotal={numericTotal}
+                selectedCells={selectedCells}
+            />
         }
     </div>
 
+}
+
+
+function BottomRow(props: {
+    selectedCells: string[],
+    header: string | { formatter: HeaderFormatter },
+    numericTotal: number
+}) {
+
+    function format(v: string | number) {
+        return typeof props.header === 'string' ? v.toString() : props.header.formatter(v)
+    }
+
+    let average = toFixedFloat(props.numericTotal / props.selectedCells.length, 2);
+
+    return <div className={`vieolo-table-interactive__bottom-row`}>
+        {
+            (props.selectedCells.length > 0 && !isNaN(props.numericTotal) && props.header) &&
+            <Flex direction='row' alignItems='center' className='height--pc-100 padding-horizontal--one' justifyContent='space-between'>
+                <TypographyParagraphSmall text={`Count: ${props.selectedCells.length}`} fontWeight='bold' />
+                <TypographyParagraphSmall text={`Average: ${format(average)}`} fontWeight='bold' />
+                <TypographyParagraphSmall text={`Total: ${format(props.numericTotal)}`} fontWeight='bold' />
+            </Flex>
+
+        }
+
+    </div>
+}
+
+
+function InputCell(props: {
+    value: string,
+    onChange: (v: string) => void,
+    onSubmit: () => void,
+    cellClass: string,
+    onCancel: () => void
+}) {
+    return <div className={`${props.cellClass} ${props.cellClass.split(" ")[0] + "--input"}`}>
+        <form onSubmit={e => {
+            e.preventDefault();
+            e.stopPropagation();
+            props.onSubmit();
+        }}>
+            <Flex alignItems='center' columnGap='half' justifyContent='space-between'>
+                <Input
+                    value={props.value}
+                    error={false}
+                    onChange={v => props.onChange(v)}
+                    size='full'
+                    autoFocus
+                    onKeyDown={e => {
+                        if (e.key === 'Escape') {
+                            props.onCancel()
+                        }
+                    }}
+                />
+                <IconButton
+                    icon={<CloseIcon />}
+                    size={'extra-small'}
+                    color={'error'}
+                    emphasis={'none'}
+                    onClick={props.onCancel}
+                    type='button'
+                />
+            </Flex>
+        </form>
+    </div>
 }